requirements.txt
python-dotenv
shiny
openai
chatlas @ git+https://github.com/posit-dev/chatlas
